requirements.txt
aiohttp==4.0.0a1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-timeout==3.0.1
attrs==23.2.0
bleach==6.1.0
caafe==0.1.5
cachetools==5.3.3
catboost==1.2.5
certifi==2024.6.2
cffi==1.17.0rc1
chardet==3.0.4
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
fonttools==4.53.0
future==1.0.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
graphviz==0.20.3
groq==0.5.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperopt==0.2.7
idna==3.7
importlib_resources==6.4.0
ipython==8.18.1
jedi==0.19.1
joblib==1.4.2
kaggle==1.6.14
kiwisolver==1.4.5
liac-arff==2.5.0
lightgbm==4.3.0
matplotlib==3.9.0
matplotlib-inline==0.1.7
minio==7.2.7
multidict==4.7.6
networkx==3.2.1
numpy==1.26.4
openai==0.28.0
openml==0.12.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pip==24.0
plotly==5.22.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
pyspark==3.5.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2rc1
regex==2024.5.15
requests==2.32.3
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
setuptools==58.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tabpfn==0.1.10
tenacity==8.3.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.5.2
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
xgboost==2.0.3
xmltodict==0.13.0
yarl==1.9.4
zipp==3.19.2